Implement manytomany relation in clientes and carrito

+++ src/main/java/org/example/proyecto_productos/Categorias/model/Categoria.java
@@ -18,7 +18,7 @@ public class Categoria {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer idCategoria;
+    private Long idCategoria;
 
     @NotBlank(message = "La descripción no puede estar vacía")
     @Size(max = 100, message = "La descripción no debe exceder los 100 caracteres")
